Determining Minimum Meal Freq

Identification of individual 6-23 months old children received the minimum meal frequency based on their breastfeeding status.

get_mmf(mmf_bf, mmf_nonbf)

get_mmf_bf(mmfbf_6to8, mmfbf_9to23)

get_mmf_bf_6to8(q4, q8, age)

get_mmf_bf_9to23(q4, q8, age)

get_mmf_nonbf(q4, q8, nonbf_frq, age)

get_nonbf_frq(q6bnum, q6cnum, q6dnum, q8)

Arguments

mmf_bf

to The binary variable indicates that the overall breastfed child received minimum meal frequency or not.

mmf_nonbf

The binary variable indicates that the overall non-breastfed child received minimum meal frequency or not.

mmfbf_6to8

to The binary variable indicates that the 6-8 months breastfed child received minimum meal frequency or not.

mmfbf_9to23

The binary variable indicates that the 9-23 months breastfed child received minimum meal frequency or not.

q4

The binary variable which mentioned that the child received breastfeeding in the previous day (yes = "1", no = "0").

q8

the integer variable with continuous scale indicates the number of meals (solid, semi-solid or soft foods) the child ate yesterday.

age

This parameter holds the information about child age in the month format.

nonbf_frq

the integer variables with continuous scale presents the frequency of non-Breastfed child meal including milk feeding.

q6bnum

This parameter indicates the number of infant formula feeding the children received in previous days.

q6cnum

This parameter holds information about the frequency of milk related food feeding the child received yesterday (milk from animals, such as fresh, tinned, or powdered milk)

q6dnum

This parameter presents the information about the frequency of yogurt drinks the child received in the previous day.

Value

binary variables indicate child get the minimum meal frequency (mmf =1 ) or not (mmf = 0)

Author

Nicholus Tint Zaw

Examples

# Minimum Meal Frequency

  df <- iycfData

  # breastfeeding children
  mmfbf_6to8 <- get_mmf_bf_6to8(df$child_bfyest,
                                df$child_food_freq,
                                df$calc_age_months)
  mmfbf_9to23 <- get_mmf_bf_9to23(df$child_bfyest,
                                  df$child_food_freq,
                                  df$calc_age_months)

  mmf_bf <- get_mmf_bf(mmfbf_6to8, mmfbf_9to23)

  # non-breastfeeding children
  nonbf_frq <- get_nonbf_frq(df$child_bms_freq,
                             df$child_milk_freq,
                             df$child_mproduct_freq,
                             df$child_food_freq)

  mmf_nonbf <- get_mmf_nonbf(df$child_bfyest,
                             df$child_food_freq,
                             nonbf_frq,
                             df$calc_age_months)

  # all children
  mmf <- get_mmf(mmf_bf, mmf_nonbf)