Infant and Young Children Indicators Dataset

This sample demo dataset (`iycfData`) contained 359 observations and 49 variables. The following tables explain the detailed description of each variable included in the sample dataset. The CARE International Myanmar Country program provided this demo dataset, and all the Personal Identifiable Information (PII) were excluded.

iycfData

Format

A data frame with 49 columns and 359 rows.

**Variables** | **Description** :--- | :--- *csex* | *calc_age_months* | *child_bf* | *child_eibf* | *child_eibf_hrs* | *child_eibf_days* | *chhild_addbf* | *chhild_addbf1* | *chhild_addbf2* | *chhild_addbf3* | *chhild_addbf4* | *chhild_addbf5* | *chhild_addbf888* | *child_bfyest* | *note_liquid* | *note_liquid2* | *child_vitdrop* | *child_ors* | *child_water* | *child_juice* | *child_broth* | *child_porridge* | *child_bms* | *child_bms_freq* | *child_milk* | *child_milk_freq* | *child_mproduct* | *child_mproduct_freq* | *child_liquid* | *note_solid_guide* | *info_solid_food* | *child_rice* | *child_potatoes* | *child_pumpkin* | *child_beans* | *child_leafyveg* | *child_mango* | *child_fruit* | *child_organ* | *child_beef* | *child_fish* | *child_insects* | *child_eggs* | *child_yogurt* | *child_fat* | *child_plam* | *child_sweets* | *child_condiments* | *child_food_freq* |

Source

will update later!

Examples

# explore the first 6 observations from the dataset
head(iycfData)
#>   csex child_bf child_eibf child_eibf_hrs child_eibf_days chhild_addbf
#> 1    1        1          0             NA              NA            5
#> 2    1        1          0             NA              NA            1
#> 3    1        1          0             NA              NA            5
#> 4    1        1          0             NA              NA            4
#> 5    1        1          0             NA              NA            5
#> 6    1        1          0             NA              NA            5
#>   chhild_addbf1 chhild_addbf2 chhild_addbf3 chhild_addbf4 chhild_addbf5
#> 1             0             0             0             0             1
#> 2             1             0             0             0             0
#> 3             0             0             0             0             1
#> 4             0             0             0             1             0
#> 5             0             0             0             0             1
#> 6             0             0             0             0             1
#>   chhild_addbf888 child_bfyest note_liquid note_liquid2 child_vitdrop child_ors
#> 1               0            1          NA           NA             0         0
#> 2               0            1          NA           NA             0         0
#> 3               0            1          NA           NA             0         0
#> 4               0            1          NA           NA             1         0
#> 5               0            0          NA           NA             0         0
#> 6               0            1          NA           NA             0         0
#>   child_water child_juice child_broth child_porridge child_bms child_bms_freq
#> 1           0           0           0              0         0             NA
#> 2           1           0           0              0         0             NA
#> 3           0           0           0              0         0             NA
#> 4           1           0           0              1         0             NA
#> 5           0           0           0              0         0             NA
#> 6           0           0           0              0         0             NA
#>   child_milk child_milk_freq child_mproduct child_mproduct_freq child_liquid
#> 1          0              NA              0                  NA            0
#> 2          0              NA              0                  NA            0
#> 3          0              NA              0                  NA            0
#> 4          0              NA              0                  NA            0
#> 5          0              NA              0                  NA            0
#> 6          0              NA              0                  NA            0
#>   note_solid_guide info_solid_food child_rice child_potatoes child_pumpkin
#> 1               NA              NA          0              0             0
#> 2               NA              NA          0              0             0
#> 3               NA              NA          0              0             0
#> 4               NA              NA          1              0             0
#> 5               NA              NA          0              0             0
#> 6               NA              NA          0              0             0
#>   child_beans child_leafyveg child_mango child_fruit child_organ child_beef
#> 1           0              0           0           0           0          0
#> 2           0              0           0           0           0          0
#> 3           0              0           0           0           0          0
#> 4           0              0           0           0           1          1
#> 5           0              0           0           0           0          0
#> 6           0              0           0           0           0          0
#>   child_fish child_insects child_eggs child_yogurt child_fat child_plam
#> 1          0             0          0            0         0          0
#> 2          0             0          0            0         0          0
#> 3          0             0          0            0         0          0
#> 4          0             0          0            0         0          0
#> 5          0             0          0            0         0          0
#> 6          0             0          0            0         0          0
#>   child_sweets child_condiments child_food_freq calc_age_months
#> 1            0                0               0               5
#> 2            0                0               3               4
#> 3            0                0               0               2
#> 4            0                1               3               0
#> 5            1                0               5               2
#> 6            0                0               0               2