Infant and Young Children Indicators Sample Dataset

This sample demo dataset (iycfData) contained 359 observations and 46 variables. The following tables explain the detailed description of each variable included in the sample dataset. The CARE International Myanmar Country program provided this demo dataset, and all the Personal Identifiable Information (PII) were excluded.

iycfData

Format

A data frame with 46 columns and 359 rows.

Variables	WHO Question Number	Description

csex		child sex

calc_age_months		child age in month

child_bf	Q1	ever breastfed

child_eibf	Q2	first put to the breast - immediately

child_eibf_hrs	Q2	first put to the breast - hours

child_eibf_days	Q2	first put to the breast - days

bf_2days	Q3	given anything other than breast milk, first 2-days of child age

child_bfyest	Q4	breastfed yesterday during the day or at night

bf_bottle	Q5	drink anything from a bottle with a nipple yesterday

child_water	Q6A	Plain water

child_bms	Q6B	Infant formula

child_bms_freq	Q6Bnum	Infant formula - frequency

child_milk	Q6C	Milk from animals

child_milk_freq	Q6Cnum	Milk from animals - frequency

child_milk_sweet	Q6Cswt	Milk from animals - sweet or flavored type milk

child_mproduct	Q6D	Yogurt drinks

child_mproduct_freq	Q6Dnum	Yogurt drinks - frequency

child_mproduct_sweet	Q6Dswt	Yogurt drinks - sweet or flavored type yogurt

child_chocolate	Q6E	Chocolate-flavored drinks

child_juice	Q6F	Fruit juice or fruit-flavored drinks

child_soda	Q6G	Sodas, malt drinks, sports drinks or energy drinks

child_tea	Q6H	Tea, coffee, or herbal drinks

child_tea_sweet	Q6Hswt	Tea, coffee, or herbal drinks - sweetened

child_broth	Q6I	Clear broth or clear soup

child_oth_drink	Q6J	Any other liquids

child_oth_drink_sweet	Q6Jswt	Other liquids - sweetened

child_yogurt	Q7A	Yogurt, other than yogurt drinks

child_yogurt_frq	Q7Anum	Yogurt, other than yogurt drinks - frequency

child_rice	Q7B	Porridge, bread, rice, noodles, pasta

child_pumpkin	Q7C	Pumpkin, carrots, sweet potatoes (yellow or orange inside)

child_potatoes	Q7D	Plantains, white potatoes, white yams, manioc, cassava

child_leafyveg	Q7E	Dark green leafy vegetables

child_oth_veg	Q7F	Any other vegetables

child_mango	Q7G	Ripe mangoes, ripe papayas

child_fruit	Q7H	Any other fruits

child_organ	Q7I	Liver, kidney, heart

child_processmeat	Q7J	Sausages, hot dogs, ham, bacon, salami, canned meat

child_beef	Q7K	Any other meat, such as beef, pork, lamb, goat, chicken, duck

child_eggs	Q7L	Eggs

child_fish	Q7M	Fresh fish, dried fish or shellfish

child_beans	Q7N	Beans, peas, lentils, nuts , seeds

child_cheese	Q7O	Hard or soft cheese

child_sweets	Q7P	Sweet foods such as chocolates, candies, cakes and biscuits

child_snack	Q7Q	Chips, crisps, puffs, French fries, instant noodles

child_oth_food	Q7R	Any other solid, semi-solid or soft food

child_food_freq	Q8	number of any solid, semi-solid or soft foods yesterday

Source

CARE Myanmar

Examples

# explore the first 6 observations from the dataset
head(iycfData)
#>   csex child_bf child_eibf child_eibf_hrs child_eibf_days chhild_addbf
#> 1    1        1          0             NA              NA            5
#> 2    1        1          0             NA              NA            1
#> 3    1        1          0             NA              NA            5
#> 4    1        1          0             NA              NA            4
#> 5    1        1          0             NA              NA            5
#> 6    1        1          0             NA              NA            5
#>   chhild_addbf1 chhild_addbf2 chhild_addbf3 chhild_addbf4 chhild_addbf5
#> 1             0             0             0             0             1
#> 2             1             0             0             0             0
#> 3             0             0             0             0             1
#> 4             0             0             0             1             0
#> 5             0             0             0             0             1
#> 6             0             0             0             0             1
#>   chhild_addbf888 child_bfyest note_liquid note_liquid2 child_vitdrop child_ors
#> 1               0            1          NA           NA             0         0
#> 2               0            1          NA           NA             0         0
#> 3               0            1          NA           NA             0         0
#> 4               0            1          NA           NA             1         0
#> 5               0            0          NA           NA             0         0
#> 6               0            1          NA           NA             0         0
#>   child_water child_juice child_broth child_porridge child_bms child_bms_freq
#> 1           0           0           0              0         0             NA
#> 2           1           0           0              0         0             NA
#> 3           0           0           0              0         0             NA
#> 4           1           0           0              1         0             NA
#> 5           0           0           0              0         0             NA
#> 6           0           0           0              0         0             NA
#>   child_milk child_milk_freq child_mproduct child_mproduct_freq child_liquid
#> 1          0              NA              0                  NA            0
#> 2          0              NA              0                  NA            0
#> 3          0              NA              0                  NA            0
#> 4          0              NA              0                  NA            0
#> 5          0              NA              0                  NA            0
#> 6          0              NA              0                  NA            0
#>   note_solid_guide info_solid_food child_rice child_potatoes child_pumpkin
#> 1               NA              NA          0              0             0
#> 2               NA              NA          0              0             0
#> 3               NA              NA          0              0             0
#> 4               NA              NA          1              0             0
#> 5               NA              NA          0              0             0
#> 6               NA              NA          0              0             0
#>   child_beans child_leafyveg child_mango child_fruit child_organ child_beef
#> 1           0              0           0           0           0          0
#> 2           0              0           0           0           0          0
#> 3           0              0           0           0           0          0
#> 4           0              0           0           0           1          1
#> 5           0              0           0           0           0          0
#> 6           0              0           0           0           0          0
#>   child_fish child_insects child_eggs child_yogurt child_fat child_plam
#> 1          0             0          0            0         0          0
#> 2          0             0          0            0         0          0
#> 3          0             0          0            0         0          0
#> 4          0             0          0            0         0          0
#> 5          0             0          0            0         0          0
#> 6          0             0          0            0         0          0
#>   child_sweets child_condiments child_food_freq calc_age_months
#> 1            0                0               0               5
#> 2            0                0               3               4
#> 3            0                0               0               2
#> 4            0                1               3               0
#> 5            1                0               5               2
#> 6            0                0               0               2